fix: complete rename — package names, import alias, label keys, short name
- Fix package declaration: nodegroup → nodedeployment in all 15 files
- Fix import alias: nodegroupcontroller → nodedeploymentcontroller
- Rename label keys: sei.io/nodegroup → sei.io/nodedeployment,
  sei.io/nodegroup-ordinal → sei.io/nodedeployment-ordinal
- Rename CRD short name: sng → snd
- Regenerate CRDs via controller-gen

All tests pass. Zero remaining references to old names.
