+++ internal/controller/nodedeployment/genesis.go
@@ -1,4 +1,4 @@
-package nodegroup
+package nodedeployment
 
 import (
 	"encoding/json"
