+++ internal/controller/nodedeployment/metrics.go
@@ -1,4 +1,4 @@
-package nodegroup
+package nodedeployment
 
 import (
 	"time"
