+++ internal/controller/nodedeployment/nodes_test.go
@@ -1,4 +1,4 @@
-package nodegroup
+package nodedeployment
 
 import (
 	"testing"
