+++ internal/controller/nodedeployment/status.go
@@ -1,4 +1,4 @@
-package nodegroup
+package nodedeployment
 
 import (
 	"context"
