+++ internal/controller/nodedeployment/status_test.go
@@ -1,4 +1,4 @@
-package nodegroup
+package nodedeployment
 
 import (
 	"testing"
